move button and NumberInput to components page
